starknet: filter and load data and deferred data
