+++ protocol/proto/starknet/v2/filter.proto
@@ -43,6 +43,14 @@ message EventFilter {
   //
   // Defaults to false.
   optional bool include_receipt = 6;
+  // Include the messages to L1 sent by the transaction that emitted the event.
+  //
+  // Defaults to false.
+  optional bool include_messages = 7;
+  // Include sibling events, that is events emitted by the same transaction.
+  //
+  // Defaults to false.
+  optional bool include_siblings = 8;
 }
 
 message Key { FieldElement value = 1; }
@@ -63,6 +71,10 @@ message MessageToL1Filter {
   //
   // Defaults to false.
   optional bool include_receipt = 4;
+  // Include the events of the transaction that sent the message.
+  //
+  // Defaults to false.
+  optional bool include_events = 5;
 }
 
 // Filter transactions.
